+++ src/pt/taimumangas/src/eu/kanade/tachiyomi/extension/pt/taimumangas/Filters.kt
